add sidebar to app
