+++ src/App.scss
@@ -1,0 +1,19 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+#wrapper {
+    width: 80%;
+}
+
+#app-wrapper {
+    display: flex;
+    flex-flow: row wrap;
+    width: 100%;
+}
+
+header {
+    width: 100%;
+    background-color: #ededed;
+}
