+++ src/components/Sidebar/GroupNav/GroupNav.css
@@ -1,0 +1,4 @@
+.group-nav ul {
+  display: flex;
+  margin: auto;
+  align-items: center; }
