+++ src/components/Sidebar/GroupNav/GroupNav.scss
@@ -1,0 +1,7 @@
+.group-nav{
+    ul {
+        display: flex;
+        margin: auto;
+        align-items: center;
+    }
+}
