+++ src/components/Sidebar/GroupNav/components/group.scss
@@ -1,0 +1,15 @@
+.group-list {
+    
+    a{
+        display: flex;
+        align-items: center;
+        width: 100%;
+        height: 5vh;
+        color: #fff;
+        text-decoration: none;
+    }
+
+    &:hover {
+        background-color: rgba(178, 178, 184, 0.349);
+    }
+}
