+++ src/components/Sidebar/Sidebar.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react';
+import GroupNav from './GroupNav/GroupNav'
+
+import './Sidebar.css'
+
+export class Sidebar extends Component {
+    render(){
+        return(
+            <div id="sidebar">
+                <div id='sidebar-wrapper'>
+                    <div className='search-fild'>
+                        <input></input>
+                    </div>
+                    <GroupNav />
+                </div>
+            </div>
+      );
+    }
+  }
