+++ src/routs.js
@@ -4,7 +4,7 @@ import { Route, Switch, } from 'react-router-dom';
 
 import Chat from './components/Chat/Chat';
 import Programs from './components/Programs/Programs';
-import Home from './components/Home/Home';
+import {Home} from './components/Home/Home';
 import GroupPage from './components/GroupPage/GroupPage';
 
 const Routs = () => (
